html/semantics/forms/the-label-element/label-inside-anchor.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>label element: clicking on label containing inline element placed inside &lt;a&gt; </title>
<link rel="author" title="Yu Han" href="mailto:yuzhehan@chromium.org">
<link rel="help" href="https://html.spec.whatwg.org/multipage/forms.html#the-label-element">
<link rel="help" href="https://html.spec.whatwg.org/multipage/text-level-semantics.html#the-a-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<a href="javascript:void(0)" target="_blank">
    <label for="foo"><span id="foo-span">foo</span></label>
    <input type="checkbox" name="foo" id="foo">
</a>

<script>
  const span = document.getElementById('foo-span'),
        cb =  document.getElementById('foo');
  let cb_change_fired = false;

  t1 = async_test("click on inline element inside a label that's placed inside a anchor should trigger default label behavior");

  cb.onchange = t1.step_func_done(function(e) {
    cb_change_fired = true;
    assert_true(cb.checked, "checkbox is checked");
  });

  t1.step(function() {
    span.click();
    assert_true(c1_change_fired);
  });

</script>
